+++ changelog.md
@@ -1,7 +1,8 @@
 # Changelog
 
 #### Version 1.9.1 - Unreleased
-- A custom desert map option. - NicEastvillage
+- Added a random desert map option. - NicEastvillage
+- Added a random sci-fi map option. - NicEastvillage
 - CycleHUD console command is now called on match state to hide HUD. - NicEastvillage
 - Mercy rule is now optional and adjustable. - NicEastvillage
 - The match is automatically stopped once mercy is hit. - NicEastvillage